requirements.txt
dnspython==2.2.1
requests==2.31.0
scapy==2.5.0
paramiko==3.4.0
ipinfo==5.0.1
timeout_decorator
pyopenvpn==0.0.3
google==3.0.0
googlesearch-python

beautifulsoup4==4.12.2
soupsieve==2.5
idna==3.4
cryptography==42.0.5
future==0.18.2
jupyter==1.0.0
pip==21.2.4
attrs==23.2.0
wheel==0.37.0
response==0.5.0
Pillow==10.0.1
tornado==6.4
Jinja2==3.1.3
setuptools==58.0.4
multidict==6.0.5
matplotlib==3.8.0
cachetools==5.3.3
aiohttp==3.9.5
yarl==1.9.4
aiosignal==1.3.1
frozenlist==1.4.1
certifi==2024.2.2
bcrypt==4.1.2
urllib3==1.26.18
pycparser==2.22
packaging==23.2
zipp==3.17.0
platformdirs==3.11.0
stem==1.8.2
httpcore==1.0.5
httpx==0.27.0
anyio==4.3.0